Move Actions part from App container
